+++ NEWS.md
@@ -16,8 +16,10 @@
 + Add `scale_variables` for scaling variables. (2022-12-02)
 + Add `Zeybel_2020` datasets. (2022-12-12)
 + Modify `run_ANOSIM`, `run_betadisper`, `run_distance`, `run_MANTEL`, `run_MRPP` and `run_PERMANOVA` for *SummarizedExperiment object*. (2022-12-13)
-+ Modify `impute_abundance`, `transform_abundances` and `trim_prevalence` for *SummarizedExperiment object*. (2022-12-14)
-+ Modify `trim_prevalence` for subset *SummarizedExperiment object*. (2023-1-7)
++ Modify `impute_abundance`, `transform_abundances` and `trim_prevalence` for *SummarizedExperiment* object. (2022-12-14)
++ Modify `trim_prevalence` for subset *SummarizedExperiment* object. (2023-1-7)
++ Add `import_SE` for *SummarizedExperiment* object. (2023-1-8)
++ Modify `scale_variables` for *SummarizedExperiment* object. (2023-1-8)
 
 
 # MicrobiomeAnalysis v1.0.2